pandora/infra-and-test-main/infra-and-test-main/container_files/tagger/pyproject.toml
[project]
name = "tagger"
version = "1.1.4"
description = "Semver tagging"
requires-python = ">=3.12"
dependencies = [
    "click==8.2.1",
    "semver==3.0.4",
    "python-gitlab==5.6.0",
    "gitpython==3.1.44",
    "toml==0.10.2",
    "laserbeak==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
]

[project.scripts]
tagger = "tagger.tagger:cli"

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = "src"
addopts = [
    "-ra",
    "-q",
    "--import-mode=importlib"
]
testpaths = [
    "tests"
]
python_files = [
    "test_*.py",
    "*_test.py"
]
python_functions = [
    "test_*",
    "*_test"
]
